moved all files in ats directory, updated setup.py
