+++ setup.py
@@ -17,17 +17,13 @@ setup(
     version=get_version(),
     description="Automated Testing System",
     install_requires=['numpy <= 1.16.5'],
-    package_dir={'': 'ats/src'},
-    packages=find_packages(
-        where='ats/src',
-    ),
-    scripts=['ats/src/ats/bin/atslite1.py', 'ats/src/ats/bin/atslite3.py'],
+    packages=['ats', 'ats.bin', 'ats.atsMachines', 'ats.util'],
     entry_points={
         'console_scripts': [
             'ats=ats.bin._ats:main',
             'ats3=ats.bin._ats3:main',
-            'atslite1=atslite1:main',
-            'atslite3=atslite3:main'
+            'atslite1=ats.bin.atslite1:main',
+            'atslite3=ats.bin.atslite3:main'
         ]
     }
 )
